requirements.txt
Flask
Pillow
tensorflow
requests
python-dotenv
google-generativeai